remove any type from table files
