+++ tornjak-frontend/axios.d.ts
@@ -1,0 +1,8 @@
+import { AxiosRequestConfig as OriginalAxiosRequestConfig } from "axios";
+
+declare module "axios" {
+  export interface AxiosRequestConfig extends OriginalAxiosRequestConfig {
+    // custom properties 
+    crossdomain: boolean;
+  }
+}
