+++ tornjak-frontend/package.json
@@ -76,6 +76,7 @@
     "@babel/preset-env": "^7.15.6",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.1.1",
+    "@types/carbon__icons-react": "^10.31.2",
     "@types/carbon-components-react": "^7.44.1",
     "@types/react-router-dom": "^5.3.1",
     "babel-jest": "^27.2.2",
