+++ tornjak-frontend/src/tables/table-head.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import { DataTable } from "carbon-components-react";
+import { DataTable, DataTableCustomHeaderData, DataTableCustomHeaderProps, DataTableCustomSelectionData, DataTableCustomSelectionProps, DataTableHeader, DataTableRow } from "carbon-components-react";
+import { ReactAttr, ShapeOf } from "carbon-components-react/typings/shared";
 const {
     TableHead,
     TableRow,
@@ -13,13 +14,13 @@ const {
 // getHeaderProps: getHeaderProps function from DataTable
 // returns header of the table for the specified entity
 type HeadProp = {
-    headers: any,
-    getSelectionProps: any,
-    getHeaderProps: any,
+    headers: DataTableHeader<string>[],
+    getSelectionProps: <E extends object = {}>(data?: ShapeOf<DataTableCustomSelectionData<DataTableRow<string>>, E> | undefined) => ShapeOf<DataTableCustomSelectionProps<DataTableRow<string>>, E> | ShapeOf<DataTableCustomSelectionProps<never>, E>,
+    getHeaderProps: <E extends object = ReactAttr<HTMLElement>>(data: ShapeOf<DataTableCustomHeaderData<DataTableHeader<string>>, E>) => ShapeOf<DataTableCustomHeaderProps<DataTableHeader<string>>, E>,
 }
 
 type HeadState = {
-    
+
 }
 class Head extends React.Component<HeadProp, HeadState> {
     constructor(props: HeadProp) {
@@ -32,8 +33,8 @@ class Head extends React.Component<HeadProp, HeadState> {
             <TableHead>
                 <TableRow>
                     <TableSelectAll {...this.props.getSelectionProps()} />
-                    {this.props.headers.map((header: { header: {} | null | undefined; }) => (
-                        <TableHeader key={header.header} {...this.props.getHeaderProps({ header })}>
+                    {this.props.headers.map((header: DataTableHeader<string>) => (
+                        <TableHeader {...this.props.getHeaderProps({ header })}>
                             {header.header}
                         </TableHeader>
                     ))}
